backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyotp>=2.9.0
qrcode[pil]>=7.4.2
sendgrid>=6.11.0
redis>=5.0.1
celery>=5.3.4
cryptography>=41.0.7
httpx>=0.25.2
python-decouple>=3.8